refactor(adaptive-dialog): fix typo in import statement and rename file to correct spelling
