+++ lib/sharezone_widgets/lib/adaptive_dialog.dart
@@ -8,7 +8,7 @@
 
 library adaptive_dialog;
 
-export 'src/adaptive_dialog/adapative_dialog_action.dart';
+export 'src/adaptive_dialog/adaptive_dialog_action.dart';
 export 'src/adaptive_dialog/column_actions_dialog/column_actions_dialog.dart';
 export 'src/adaptive_dialog/left_and_right_action_dialog/left_and_right_action_adaptive_dialog.dart';
 export 'src/adaptive_dialog/long_press_dialog/long_press.dart';
