+++ lib/sharezone_widgets/lib/src/adaptive_dialog/adaptive_dialog_action.dart
