+++ lib/sharezone_widgets/lib/src/adaptive_dialog/left_and_right_action_dialog/left_and_right_action_adaptive_dialog.dart
@@ -13,7 +13,7 @@ import 'package:sharezone_utils/platform.dart';
 
 import '../../../theme.dart';
 import '../../dialog_wrapper.dart';
-import '../adapative_dialog_action.dart';
+import '../adaptive_dialog_action.dart';
 
 Future<T> showLeftRightAdaptiveDialog<T>({
   @required BuildContext context,
